+++ R/gs.R
@@ -1,16 +1,18 @@
 #' Function that takes the Google Spreadsheet document key to save it in the working dir
 #'
 #' @param x key of the Google Sheet document
-#' @param extension file extension for local data, either `csv`` or `json`. defaults to csv.
-use_gs <- function(x, extension = "csv") {
+#' @param extension file extension for local data, either `csv` or `json`. defaults to csv.
+#' @export
+use_gs_acceptance <- function(x, extension = "csv") {
   path = "tests/testthat/data"
-
-  gap <- googlesheets::gs_key(x, lookup = FALSE, visibility = "private")
+  gap <- googlesheets::gs_key(
+    x,
+    lookup = FALSE,
+    visibility = "private"
+  )
   tabs <- gap$ws$ws_title
-  title <- gap$sheet_title
+  title <- clean_dir_name(gap$sheet_title)
 
-  title <- tolower(title)
-  title <- gsub(" ", "-", title) # replace spaces
   data.dir = paste0(path, "/", title, "/")
   if (!dir.exists(data.dir)) {
     dir.create(data.dir, recursive = TRUE)
@@ -19,7 +21,7 @@ use_gs <- function(x, extension = "csv") {
   lapply(tabs, function(ws) {
     dt <- googlesheets::gs_read(gap, ws)
 
-    gs_save_worksheet(
+    gs_import_worksheet(
       x = dt,
       name = ws,
       dir = data.dir,
@@ -32,7 +34,7 @@ use_gs <- function(x, extension = "csv") {
 #' @noRd
 #' @importFrom utils write.csv
 #' @import data.table
-gs_save_worksheet <- function(x, name, dir, extension) {
+gs_import_worksheet <- function(x, name, dir, extension) {
   mock_file <- NULL
   save_fn <- make_save_mock_data(extension)
 
@@ -80,4 +82,13 @@ make_save_mock_data <- function(extension) {
       save_mock_data_json
     }
   )
+}
+
+#' Conforms the name of the directory to be lower case and dash separated
+#' @noRd
+#' @param x string to turn into a folder name
+clean_dir_name <- function(x) {
+  x <- tolower(x)
+  x <- gsub(" ", "-", x) # replace spaces
+  x
 }