+++ NEWS.md
@@ -1,3 +1,3 @@
 # usegs 0.1.0
 
-* use_gs() - allows to save Google Sheet to the local files in the testthat folder.
+* use_gs_acceptance() - allows to save Google Sheet to the local files in the testthat folder.